+++ Asteroids/Assets/Resources/Audio/Effects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4158a46be6261f94787dee71651289c1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
