+++ Asteroids/Assets/Scenes/MainMenu.unity
@@ -245,10 +245,73 @@ Transform:
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 1, z: -10}
   m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
+  m_Children:
+  - {fileID: 1645754356}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1001 &1645754355
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 1607419809}
+    m_Modifications:
+    - target: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1184579799394220167, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+      propertyPath: m_Name
+      value: AudioSources
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+--- !u!4 &1645754356 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 1184579799394220166, guid: ac50d10fe686b554593963f19d275b93, type: 3}
+  m_PrefabInstance: {fileID: 1645754355}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1 &2028356340
 GameObject:
   m_ObjectHideFlags: 0
