+++ Asteroids/Assets/Sources/FactoriesAndPools.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9738149219bf89d48b2e9ae80512f653
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
