+++ Asteroids/Assets/Sources/ResourceManagers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8db988b516055d342b6280240a6cedf0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
