+++ Asteroids/Assets/Sources/ResourceManagers/ResourceManager.cs.meta
