+++ Asteroids/Assets/Sources/ResourceManagers/ResourceManagerSettings.cs.meta
