+++ Asteroids/Assets/Sources/Starters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2a9c21b809aa1ff4face7d7c585c67d3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
