+++ Asteroids/Assets/Sources/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1290dea31e45ccf46b5a057eeace3ea5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
